Adding support for Cedar stack so that scale is used to add workers.
The original resque-heroku-autoscaler does NOT work yet with the new Heroku's
Cedar stack. In particular the heroku_client.set_workers() will fail on Cedar
with the error:
        422 Unprocessable Entity: <errors><error>For Cedar apps, use `heroku scale worker 1`</error></errors>

That is because set_workers() corresponds to the heroku-cli "heroku workers <qty>"
and on cedar the required command is "heroku scale workers=<qty>".
On this gist https://gist.github.com/1082673 there's a patch to HireFire that shows
how to scale using the heroku gem's ps_scale() method.

In this revision I've added a current_stack() method that returns the
name of the current stack, and them modified set_workers() to check for
the current stack and use ps_scale() in case of cedar.