BUG FIX and bumped up version for deploy.
Was inconsistently using 'worker' and 'workers' to refer to the worker
processes instantiated on heroku.
This caused a problem because e.g. a scaled up 'worker' would not match
the current_workers count in set_workers, and no auto down-scaling was
performerd.
In other words, starting from a no worker status, enqueuing a job would
case the instantiation of a 'worker' job. Once it is finished and the
after_perform_scale_workers() is executed, the scale-to-zero is executed
but the check 'if number_of_workers != current_workers' in set_workers()
would be false because current_workers returns the processes named
'workers', and there is none because the worker instantiated was named
'worker'.

FIXED by using 'workers' in the ps_scale() call.